+++ src/ADAMS_solveLocalBS.h
@@ -12,12 +12,12 @@ namespace ADAMS{
   const int AMO=7; //must be between 5 and 8 (for now). 7 Seems good.
 
   int solveDBS(std::vector<double> &p, std::vector<double> &q, double &en,
-      std::vector<double> v, int Z, int n, int ka,
+      std::vector<double> v, double Z, int n, int ka,
       std::vector<double> r, std::vector<double> drdt, double h, int NGP,
       int &pinf, int &its, double &eps, double alpha, int log_dele_or=0);
 
   int outwardAM(std::vector<double> &p, std::vector<double> &q, double &en,
-      std::vector<double> v, int Z, int ka,
+      std::vector<double> v, double Z, int ka,
       std::vector<double> r, std::vector<double> drdt, double h, int NGP,
       int ctp, double alpha);
