+++ src/ADAMS_solveLocalContinuum.h
@@ -8,7 +8,7 @@ double fitQuadratic(double x1, double x2, double x3, double y1, double y2,
   double y3);
 
 int solveContinuum(std::vector<double> &p, std::vector<double> &q, double en,
-    std::vector<double> v, int Z, int ka,
+    std::vector<double> v, double Z, int ka,
     std::vector<double> r, std::vector<double> drdt, double h, int NGPb, int NGPc, int i_asym,
     double alpha);
